Finished My Level
Okay. So I made a new scene with a BUNCH of new prefabs for platforms that move. I was thinking that we could each make a short level with three money in each so we don't have to mess with the scripts too much. There are moving versions of each platform and wall versions. A couple of things I noticed: the bananas need to be deleted if they hit any type of platform or their specific color enemy. When I tried to put in enemies, it kept infinitely spawning enemies from the array, so I just put sprites where I want enemies in my level. We also need a way to link the levels together upon beating the previous, and along with that, the tutorial should only be on the first spawn in for the player. Also, I'm pretty sure I fixed the finish spawn problem. If we just keep it out of any autogeneration scripts it shouldn't get messed up.

Okay, I think that's everything